+++ cache.go
@@ -3,6 +3,7 @@ package main
 import (
 	"fmt"
 	"log"
+
 	"github.com/VictoriaMetrics/fastcache"
 	"github.com/labstack/echo"
 )
